+++ web-app/src/app/app.module.ts
@@ -157,6 +157,8 @@ import { AdminAuthenticationSAMLComponent } from './admin/admin-authentication/a
 import { ButtonPreviewComponent } from './admin/admin-authentication/admin-authentication-create/button-preview/button-preview.component';
 import { AdminAuthenticationSettingsComponent } from './admin/admin-authentication/admin-authentication-settings.component';
 import { AdminSettingsUnsavedComponent } from './admin/admin-settings/admin-settings-unsaved/admin-settings-unsaved.component';
+import { AdminEventFormPreviewComponent } from './admin/admin-event/admin-event-form/admin-event-form-preview/admin-event-form-preview.component';
+import { AdminEventFormPreviewDialogComponent } from './admin/admin-event/admin-event-form/admin-event-form-preview/admin-event-form-preview-dialog.component';
 
 @NgModule({
   declarations: [
@@ -241,7 +243,9 @@ import { AdminSettingsUnsavedComponent } from './admin/admin-settings/admin-sett
     AdminAuthenticationSAMLComponent,
     ButtonPreviewComponent,
     AdminAuthenticationSettingsComponent,
-    AdminSettingsUnsavedComponent
+    AdminSettingsUnsavedComponent,
+    AdminEventFormPreviewComponent,
+    AdminEventFormPreviewDialogComponent
   ],
   imports: [
     CommonModule,