+++ CHANGELOG.md
@@ -17,6 +17,7 @@ MAGE adheres to [Semantic Versioning](http://semver.org/).
 * Fix observation view error on invalid primary/secondary feed field.
 * Preserve primary and secondary feed fields on form import.
 * Fix feed item preview, account for feed item without attachments.
+* Fix form preview on event admin page.
 
 ## [6.0.1](https://github.com/ngageoint/mage-server/releases/tag/6.0.1)